--Move trajectory tube synthesizing code to esp::geo;Fix winding.
