+++ src/esp/assets/ResourceManager.cpp
@@ -1206,192 +1206,6 @@ scene::SceneNode* ResourceManager::createRenderAssetInstanceGeneralPrimitive(
   return &newNode;
 }
 
-std::vector<Mn::Vector3> ResourceManager::buildSmoothTrajOfPoints(
-    const std::vector<Mn::Vector3>& pts,
-    int numInterp) {
-  std::vector<Mn::Math::CubicHermite<Mn::Vector3>> splinePath;
-  std::vector<Mn::Vector3> trajectory;
-  Mn::Vector3 a, b;
-  // pts.size() must be > 1
-  // beginning point - tangents are opposites
-  splinePath.emplace_back(buildSpline({(pts[0] - pts[1]).normalized()},
-                                      {(pts[1] - pts[0]).normalized()},
-                                      pts[0]));
-  int numPtsM1 = pts.size() - 1;
-  for (int i = 1; i < numPtsM1; ++i) {
-    splinePath.emplace_back(buildSpline({(pts[i] - pts[i - 1]).normalized()},
-                                        {(pts[i + 1] - pts[i]).normalized()},
-                                        pts[i]));
-
-    for (int j = 0; j < numInterp; ++j) {
-      double t = j / (1.0 * numInterp);
-      trajectory.emplace_back(
-          Mn::Math::splerp(splinePath[i - 1], splinePath[i], t));
-    }
-  }
-
-  // end point - tangents are opposites
-  splinePath.emplace_back(buildSpline(
-      {(pts[numPtsM1] - pts[numPtsM1 - 1]).normalized()},
-      {(pts[numPtsM1 - 1] - pts[numPtsM1]).normalized()}, pts[numPtsM1]));
-  for (int j = 0; j < numInterp; ++j) {
-    double t = j / (1.0 * numInterp);
-    trajectory.emplace_back(
-        Mn::Math::splerp(splinePath[numPtsM1 - 1], splinePath[numPtsM1], t));
-  }
-  return trajectory;
-}  // ResourceManager::buildSmoothTrajOfPoints
-
-Mn::Trade::MeshData ResourceManager::trajectoryTubeSolid(
-    const std::vector<Mn::Vector3>& pts,
-    int numSegments,
-    float radius,
-    bool smooth,
-    int numInterp) {
-  // 1. Build smoothed trajectory through passed points if requested
-  std::vector<Mn::Vector3> trajectory =
-      (smooth ? buildSmoothTrajOfPoints(pts, numInterp) : pts);
-
-  // 2. Build mesh vertex points around each trajectory point at appropriate
-  // distance (radius). For each point in trajectory, add a wireframe circle
-  // centered at that point, appropriately oriented based on tangents
-
-  // make sure importer is open before use
-  primitiveImporter_->openData("");
-  // configuration for PrimitiveImporter - replace appropriate group's data
-  // before instancing prim object
-  Cr::Utility::ConfigurationGroup* cfgGroup =
-      primitiveImporter_->configuration().group("circle3DWireframe");
-  if (cfgGroup != nullptr) {
-    // set value for config group for 3D circle wireframes
-    cfgGroup->setValue<int>("segments", numSegments);
-  }
-
-  // get verts for circle primitive to use as endpoints and transform them to
-  // more easily use them to calculate tangents at appropriate raidius and
-  // normals
-  Corrade::Containers::Array<Magnum::Vector3> circleVerts =
-      primitiveImporter_->mesh("circle3DWireframe")->positions3DAsArray();
-  ASSERT(numSegments == circleVerts.size());
-
-  // normalized verts
-  Corrade::Containers::Array<Magnum::Vector3> circleNormVerts{
-      Cr::Containers::NoInit, sizeof(Magnum::Vector3) * numSegments};
-  // transform points to be on circle of given radius, and make copy to
-  // normalize points
-  for (int i = 0; i < numSegments; ++i) {
-    circleVerts[i] *= radius;
-    circleNormVerts[i] = circleVerts[i].normalized();
-  }
-  // size of trajectory
-  const Mn::UnsignedInt trajSize = trajectory.size();
-  // # of vertices in resultant tube == # circle verts * # points in trajectory
-  const Mn::UnsignedInt vertexCount = numSegments * trajSize;
-  struct Vertex {  // a function-local struct
-    Mn::Vector3 position;
-    Mn::Vector3 normal;
-  };
-  // Vertex data storage
-  Cr::Containers::Array<char> vertexData{Cr::Containers::NoInit,
-                                         sizeof(Vertex) * vertexCount};
-
-  Cr::Containers::StridedArrayView1D<Vertex> vertices =
-      Cr::Containers::arrayCast<Vertex>(vertexData);
-  // Position and normal views of vertex array
-  Cr::Containers::StridedArrayView1D<Mn::Vector3> positions =
-      vertices.slice(&Vertex::position);
-  Cr::Containers::StridedArrayView1D<Mn::Vector3> normals =
-      vertices.slice(&Vertex::normal);
-
-  Mn::UnsignedInt circlePtIDX = 0;
-  for (Mn::UnsignedInt vertIx = 0; vertIx < trajSize; ++vertIx) {
-    const Mn::Vector3& vert = trajectory[vertIx];
-    Mn::Vector3 tangent;
-    if (vertIx == 0) {  // first vert
-      tangent = trajectory[1] - trajectory[0];
-    } else if (vertIx == trajSize - 1) {  // last vert
-      tangent = trajectory[trajSize - 1] - trajectory[trajSize - 2];
-    } else {  // other verts, use tangent average
-      Mn::Vector3 pTangent = trajectory[vertIx] - trajectory[vertIx - 1];
-      Mn::Vector3 nTangent = trajectory[vertIx + 1] - trajectory[vertIx];
-      tangent = (pTangent + nTangent) / 2.0;
-    }
-    // get the orientation matrix assuming y-up preference
-    Mn::Matrix4 tangentOrientation =
-        Mn::Matrix4::lookAt(vert, vert + tangent, Mn::Vector3{0, 1.0, 0});
-    for (int i = 0; i < numSegments; ++i) {
-      // build vertex (circleVerts[i] is at radius)
-      positions[circlePtIDX] =
-          tangentOrientation.transformPoint(circleVerts[i]);
-      // pre-rotated normal for circle is normalized point
-      normals[circlePtIDX] =
-          tangentOrientation.transformVector(circleNormVerts[i]);
-      ++circlePtIDX;
-    }
-  }
-
-  // 3. Create polys between all points
-  Cr::Containers::Array<char> indexData{
-      Cr::Containers::NoInit,
-      6 * numSegments * (trajSize - 1) * sizeof(Mn::UnsignedInt)};
-  Cr::Containers::ArrayView<Mn::UnsignedInt> indices =
-      Cr::Containers::arrayCast<Mn::UnsignedInt>(indexData);
-
-  // create triangle indices for each tube pair correspondance - ccw winding
-  /*
-            +n---+n+1
-            | \ F2|
-            |  \  |
-            |F1 \ |
-            +0---+1
-        F1 = [+0, +1, +n]
-        F2 = [+1, +n+1, +n]
-   */
-  int iListIDX = 0;
-  for (Mn::UnsignedInt vIdx = 0; vIdx < trajSize - 1;
-       ++vIdx) {  // skip last circle (adding forward)
-    int vIdxNumSeg = vIdx * numSegments;
-    for (Mn::UnsignedInt circleIx = 0; circleIx < numSegments; ++circleIx) {
-      Mn::UnsignedInt ix = circleIx + vIdxNumSeg;  //+0
-      Mn::UnsignedInt ixNext = ix + numSegments;   //+n
-      Mn::UnsignedInt ixPlus = ix + 1;             //+1
-      Mn::UnsignedInt ixNextPlus = ixNext + 1;     //+n+1
-      if (circleIx == numSegments - 1) {
-        // last vert in a circle wraps to relative 0
-        ixPlus = vIdxNumSeg;
-        ixNextPlus = vIdxNumSeg + numSegments;
-      }
-      // F1
-      indices[iListIDX++] = (ix);
-      indices[iListIDX++] = (ixPlus);
-      indices[iListIDX++] = (ixNext);
-      // F2
-      indices[iListIDX++] = (ixPlus);
-      indices[iListIDX++] = (ixNextPlus);
-      indices[iListIDX++] = (ixNext);
-    }
-  }
-
-  // Finally, make the MeshData. The indices have to be constructed first
-  // because function argument evaluation order is not guaranteed and so you
-  // might end up with the move happening before the MeshIndexData construction,
-  // which would result in 0 indices)
-
-  // Building mesh this way should obviate the need for interleaving
-
-  Mn::Trade::MeshData meshData{
-      Mn::MeshPrimitive::Triangles,
-      std::move(indexData),
-      Mn::Trade::MeshIndexData{indices},
-      std::move(vertexData),
-      {Mn::Trade::MeshAttributeData{Mn::Trade::MeshAttribute::Position,
-                                    positions},
-       Mn::Trade::MeshAttributeData{Mn::Trade::MeshAttribute::Normal, normals}},
-      static_cast<Mn::UnsignedInt>(positions.size())};
-
-  return meshData;
-}  // ResourceManager::trajectoryTubeSolid
-
 bool ResourceManager::loadTrajectoryVisualization(
     const std::string& trajVisName,
     const std::vector<Mn::Vector3>& pts,
@@ -1417,9 +1231,32 @@ bool ResourceManager::loadTrajectoryVisualization(
             << numSegments << " circular segments and " << numInterp
             << " interpolated points between each trajectory point.";
 
+  // get circular vertices to use to construct the trajectory tube
+  // make sure importer is open before use
+  primitiveImporter_->openData("");
+  // configuration for PrimitiveImporter - replace appropriate group's data
+  // before instancing prim object
+  Cr::Utility::ConfigurationGroup* cfgGroup =
+      primitiveImporter_->configuration().group("circle3DWireframe");
+  CORRADE_ASSERT(cfgGroup != nullptr,
+                 "ResourceManager::loadTrajectoryVisualization: Error with "
+                 "primitiveImporter_ : Failed to supply circle3DWireframe "
+                 "configuration. Aborting.",
+                 false);
+
+  // set value for config group for 3D circle wireframes
+  cfgGroup->setValue<int>("segments", numSegments);
+
+  // get verts for circle primitive to use as endpoints and transform them to
+  // more easily use them to calculate tangents at appropriate raidius and
+  // normals
+  Cr::Containers::Array<Magnum::Vector3> circleVerts =
+      primitiveImporter_->mesh("circle3DWireframe")->positions3DAsArray();
+  ASSERT(numSegments == circleVerts.size());
+
   // create mesh tube
   Cr::Containers::Optional<Mn::Trade::MeshData> trajTubeMesh =
-      trajectoryTubeSolid(pts, numSegments, radius, smooth, numInterp);
+      geo::trajectoryTubeSolid(pts, circleVerts, radius, smooth, numInterp);
   LOG(INFO) << "ResourceManager::loadTrajectoryVisualization : Successfully "
                "returned from trajectoryTubeSolid ";
